pyproject.toml
[tool.poetry]
name = "cataloguer"
version = "2.0"
description = "Organize your media files"
authors = ["Iago Veloso"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/iago1460/cataloguer"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.9"
Pillow = "8.1.2"
python-dateutil = "2.8.2"
python-magic = "0.4.27"
#progressbar2 = "3.53.1"
pydantic = "^1.10.5"
click = "^8.1.3"
rich-click = "^1.6.1"


[tool.poetry.dev-dependencies]
pytest = "~7.2.1"
pytest-mock = "~3.10"

[tool.poetry.scripts]
cataloguer = "cataloguer.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
